+++ clodoo/manage_2_dbs.py
@@ -1,7 +1,15 @@
 #!/home/odoo/devel/venv/bin/python2
 # -*- coding: utf-8 -*-
 from __future__ import print_function, unicode_literals
+from __future__ import absolute_import
+from __future__ import division
 
+from future import standard_library
+standard_library.install_aliases()                                 # noqa: E402
+# from builtins import str
+# from builtins import input
+# from builtins import range
+# from builtins import *                                           # noqa: F403
 import sys
 import time
 # import oerplib
@@ -96,7 +104,7 @@ def writelog(msg):
 def manage_error():
     dummy = ''
     while dummy not in ('I', 'i', 'S', 's', 'D', 'd'):
-        dummy = raw_input('(Ignore, Stop, Debug)? ')
+        dummy = input('(Ignore, Stop, Debug)? ')
         if not dummy:
             dummy = 'I'
         if dummy == 'S' or dummy == 's':
@@ -164,7 +172,6 @@ def write_no_dup(ctx, model, ids, vals, rec_id):
         except BaseException:
             writelog('Error writing record %d of %s' % (rec_id, model))
             manage_error()
-            pass
 
 
 def create_with_id(ctx, model, id, vals):
@@ -225,7 +232,7 @@ def create_with_id(ctx, model, id, vals):
     if new_id != id:
         writelog("Cannot create record %d of %s" % (id, model))
         if not ctx['assume_yes']:
-            raw_input('Press RET to continue')
+            input('Press RET to continue')
     if last_id and last_id > 1 and last_id > id:
         sql = 'alter sequence %s restart %d;' % (table, last_id)
         try:
@@ -233,16 +240,17 @@ def create_with_id(ctx, model, id, vals):
         except BaseException:
             pass
 
+
 def install_modules(dst_ctx, src_ctx):
     assume_yes = dst_ctx['assume_yes']
     upgrade = False
     if not dst_ctx['assume_yes']:
-        dummy = raw_input('Install modules (Yes,No,All)? ')
+        dummy = input('Install modules (Yes,No,All)? ')
         if dummy[0] in ('n', 'N'):
             assume_yes = 'N'
         if dummy[0] in ('a', 'A'):
             assume_yes = 'Y'
-        dummy = raw_input('Upgrade installed modules (Yes,No)? ')
+        dummy = input('Upgrade installed modules (Yes,No)? ')
         if dummy[0] in ('y', 'Y'):
             upgrade = True
         if assume_yes == 'N' and not upgrade:
@@ -273,7 +281,7 @@ def install_modules(dst_ctx, src_ctx):
                 dummy = 'Y'
             else:
                 if not assume_yes:
-                    dummy = raw_input(
+                    dummy = input(
                         'Install module %s (Yes,No,All,Quit)? ' % module)
             if dummy[0] in ('n', 'N'):
                 continue
@@ -360,7 +368,7 @@ def cvt_o2m_value(dst_ctx, src_ctx, model, name, value, format=False):
                                     mode=rel_mode)
                     else:
                         writelog('Model %s id %d does not exits!' % (model,
-                                                                    id))
+                                                                     id))
                         src_ctx['_CACHE'][model][id] = False
         value = new_value if new_value else False
     if format == 'cmd' and value:
@@ -408,7 +416,7 @@ def cvt_m2m_value(dst_ctx, src_ctx, model, name, value, format=False):
                                     mode=rel_mode)
                     else:
                         writelog('Model %s id %d does not exits!' % (model,
-                                                                    id))
+                                                                     id))
                         src_ctx['_CACHE'][model][id] = False
         value = new_value if new_value else False
     if format == 'cmd' and value:
@@ -455,7 +463,7 @@ def cvt_m2o_value(dst_ctx, src_ctx, model, name, id, format=False):
                                 mode=rel_mode)
                 else:
                     writelog('Model %s id %d does not exits!' % (model,
-                                                                id))
+                                                                 id))
                     src_ctx['_CACHE'][model][id] = False
         id = new_id
     return id
@@ -466,7 +474,6 @@ def cvt_state_value(dst_ctx, src_ctx, model, name, value):
         value = 'draft'
     elif value == 'paid':
         value = 'draft'
-    pass
 
 
 def load_record(dst_ctx, src_ctx, model, rec, mode=None):
@@ -516,7 +523,7 @@ def copy_record(dst_ctx, src_ctx, model, rec, mode=None):
         ids = clodoo.searchL8(dst_ctx, model,
                               where)
         if not ids and hasattr(rec, 'active'):
-            where.append(('active','=',False))
+            where.append(('active', '=', False))
             ids = clodoo.searchL8(dst_ctx, model,
                                   where)
         if len(ids) > 1:
@@ -530,7 +537,7 @@ def copy_record(dst_ctx, src_ctx, model, rec, mode=None):
             except:
                 writelog('Cannot create %s src id=%d' % (model, rec.id))
                 manage_error()
-                pass
+
 
 def copy_table(dst_ctx, src_ctx, model, mode=None):
     clodoo.get_model_structure(src_ctx, model,
@@ -672,11 +679,14 @@ def build_table_tree(ctx):
             models[model]['level'] = MAX_DEEP + 1
     return models
 
+
 PKEYS = {
     'res.country.state': ['country_id', 'code'],
     'res.partner': ['name', 'vat'],
     'res.company': ['vat'],
 }
+
+
 def primkey_table(ctx, model):
     clodoo.get_model_structure(ctx, model,
                                ignore=IGNORE_FIELDS.get(model, []))
@@ -747,7 +757,7 @@ def get_model_copy_mode(ctx, model):
         mode_selection = {'i': 'image', 's': 'sql', 'n': 'no'}
         dummy = ''
         while not dummy:
-            dummy = raw_input('Copy table %s (Image,Sql,No)? ' % model)
+            dummy = input('Copy table %s (Image,Sql,No)? ' % model)
             if dummy and dummy[0].lower() in mode_selection:
                 mode = mode_selection[dummy[0].lower()]
             else: