+++ docs/conf.py
@@ -25,9 +25,9 @@ copyright = '2019-21, SHS-AV s.r.l.'
 author = 'Antonio Maria Vigliotti'
 
 # The short X.Y version
-version = '1.0.5.46'
+version = '1.0.5.47'
 # The full version, including alpha/beta/rc tags
-release = '1.0.5.46'
+release = '1.0.5.47'
 
 
 # -- General configuration ---------------------------------------------------