+++ z0bug_odoo/__main__.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/home/odoo/devel/venv/bin/python2
 # -*- coding: utf-8 -*-
 """z0bug_odoo
 """
@@ -20,7 +20,7 @@ except ImportError:
 standard_library.install_aliases()
 
 
-__version__ = "1.0.5"
+__version__ = "1.0.5.1"
 STS_FAILED = 1
 STS_SUCCESS = 0