+++ z0bug_odoo/create_test_env.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/home/odoo/devel/venv/bin/python2
 # -*- coding: utf-8 -*-
 """
 Create Test Environment base
@@ -38,7 +38,7 @@ except ImportError:
     from z0lib import z0lib
 from z0bug_odoo import z0bug_odoo_lib
 
-__version__ = "1.0.5"
+__version__ = "1.0.5.1"
 
 
 def env_ref(ctx, xref, retxref_id=None):