+++ z0bug_odoo/docs/conf.py
@@ -12,22 +12,22 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+# sys.path.insert(0, os.path.abspath('../'))
 # import sphinx_rtd_theme
 
 
 # -- Project information -----------------------------------------------------
 
 project = 'z0bug_odoo'
-copyright = '2019-20, SHS-AV s.r.l.'
+copyright = '2019-21, SHS-AV s.r.l.'
 author = 'Antonio Maria Vigliotti'
 
 # The short X.Y version
-version = '1.0.5'
+version = '1.0.5.1'
 # The full version, including alpha/beta/rc tags
-release = '1.0.5'
+release = '1.0.5.1'
 
 
 # -- General configuration ---------------------------------------------------
@@ -44,6 +44,7 @@ extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.githubpages',
     'sphinx.ext.autodoc',
+    'sphinx.ext.autosummary',
 ]
 
 
@@ -137,3 +138,11 @@ html_static_path = ['_static']
 #
 # html_sidebars = {}
 html_logo = 'logozero_180x46.png'
+#
+# autodoc_default_flags = ['members']
+# autosummary_generate = True
+# autodoc_default_options = {
+#     'members': 'Z0test, Z0testOdoo',
+#     'undoc-members': True,
+#     'exclude-members': 'Macro, SanityTest'
+# }