+++ z0bug_odoo/tests/z0bug_odoo_test_02.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 # -*- coding: utf-8 -*-
 # Copyright (C) 2015-2020 SHS-AV s.r.l. (<http://www.zeroincombenze.org>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
@@ -14,7 +14,7 @@ import shutil
 # from z0bug_odoo import test_common
 from zerobug import Z0testOdoo
 
-__version__ = "1.0.5"
+__version__ = "1.0.5.1"
 
 MODULE_ID = 'z0bug_odoo'
 TEST_FAILED = 1